payments/pyproject.toml
[project]
name = "sveltereader-payments"
version = "0.0.1"
description = "Cashu payments service for SvelteReader"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0,<0.115.0",
    "uvicorn[standard]",
    "python-dotenv",
    "pydantic>=1.10.2,<2.0.0",
    "marshmallow>=3.13,<4.0",
    "cashu>=0.18.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.24.0", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.ruff]
lint.select = ["E", "F", "I", "UP"]
lint.ignore = ["UP006", "UP007", "UP035", "E501"]
